Split out `BaseSession` from `DocSession`.
