+++ local-modules/@bayou/doc-server/DocSession.js
@@ -2,13 +2,10 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { Storage } from '@bayou/config-server';
-import { BodyChange, CaretId, PropertyChange } from '@bayou/doc-common';
+import { BodyChange, PropertyChange } from '@bayou/doc-common';
 import { RevisionNumber, Timestamp } from '@bayou/ot-common';
-import { TBoolean } from '@bayou/typecheck';
-import { CommonBase, Errors } from '@bayou/util-common';
 
-import FileComplex from './FileComplex';
+import BaseSession from './BaseSession';
 
 /**
  * Server side representative of an editing session for a specific document,
@@ -19,7 +16,7 @@ import FileComplex from './FileComplex';
  * underlying `*Control` instances, while implicitly adding author ID and/or
  * caret ID as appropriate to methods that perform modifications.
  */
-export default class DocSession extends CommonBase {
+export default class DocSession extends BaseSession {
   /**
    * Constructs an instance.
    *
@@ -32,36 +29,9 @@ export default class DocSession extends CommonBase {
    *   view-only session.
    */
   constructor(fileComplex, authorId, caretId, canEdit) {
-    super();
+    super(fileComplex, authorId, caretId, canEdit);
 
-    /** {FileComplex} File complex that this instance is part of. */
-    this._fileComplex = FileComplex.check(fileComplex);
-
-    /** {string} Author ID. */
-    this._authorId = Storage.dataStore.checkAuthorIdSyntax(authorId);
-
-    /** {string} Caret ID. */
-    this._caretId = CaretId.check(caretId);
-
-    /** {boolean} Whether or not this instance allows edits. */
-    this._canEdit = TBoolean.check(canEdit);
-
-    // **TODO:** Remove this restriction!
-    if (!canEdit) {
-      throw Errors.wtf('View-only sessions not yet supported!');
-    }
-
-    /** {BodyControl} The underlying body content controller. */
-    this._bodyControl = fileComplex.bodyControl;
-
-    /** {CaretControl} The underlying caret info controller. */
-    this._caretControl = fileComplex.caretControl;
-
-    /** {PropertyControl} The underlying property (metadata) controller. */
-    this._propertyControl = fileComplex.propertyControl;
-
-    /** {Logger} Logger to use to relay events coming from the client. */
-    this._clientLog = this._fileComplex.fileAccess.log.withAddedContext('client');
+    Object.freeze(this);
   }
 
   /**
@@ -289,93 +259,5 @@ export default class DocSession extends CommonBase {
     const change = new PropertyChange(baseRevNum + 1, delta, Timestamp.now(), this._authorId);
 
     return this._propertyControl.update(change);
-  }
-
-  /**
-   * Indicates whether (`true`) or not (`false`) this instance allows editing to
-   * be performed through it.
-   *
-   * **Note:** This is a method and not just a property, specifically so that
-   * clients (via the API) can make this determination.
-   *
-   * @returns {boolean} `true` if this instance allows editing, or `false` if it
-   *   is view-only.
-   */
-  canEdit() {
-    return this._canEdit;
-  }
-
-  /**
-   * Returns a bit of identifying info about this instance, for the purposes of
-   * logging. Specifically, the client will call this method and log the result
-   * during session initiation.
-   *
-   * @returns {object} Succinct identification.
-   */
-  getLogInfo() {
-    const result = {
-      authorId:   this.getAuthorId(),
-      caretId:    this.getCaretId(),
-      documentId: this.getDocumentId(),
-      fileId:     this.getFileId(),
-      canEdit:    this.canEdit()
-    };
-
-    // Only include the file ID if it's not the same as the document ID.
-    if (result.fileId === result.documentId) {
-      delete result.fileId;
-    }
-
-    return result;
-  }
-
-  /**
-   * Returns the author ID of the user managed by this session.
-   *
-   * @returns {string} The author ID.
-   */
-  getAuthorId() {
-    return this._authorId;
-  }
-
-  /**
-   * Returns the caret ID of this instance.
-   *
-   * @returns {string} The caret ID.
-   */
-  getCaretId() {
-    return this._caretId;
-  }
-
-  /**
-   * Returns the ID of the document controlled by this instance.
-   *
-   * @returns {string} The document ID.
-   */
-  getDocumentId() {
-    return this._fileComplex.fileAccess.documentId;
-  }
-
-  /**
-   * Returns the ID of the file controlled by this instance.
-   *
-   * @returns {string} The file ID.
-   */
-  getFileId() {
-    return this._fileComplex.fileAccess.file.id;
-  }
-
-  /**
-   * Causes an event (which will come from the client) to be logged here on the
-   * server. This is useful for tactical debugging, moreso than intended for
-   * long-term use.
-   *
-   * **TODO:** Consider removing this.
-   *
-   * @param {string} name The event name to log.
-   * @param {...*} args Arbitrary arguments to log.
-   */
-  logEvent(name, ...args) {
-    this._clientLog.event[name](...args);
   }
 }
